move accept invite logic to separate controller
