+++ app/controllers/accept_invites_controller.rb
@@ -1,0 +1,42 @@
+class AcceptInvitesController < ApplicationController
+  layout 'devise'
+
+  def new
+    @token = params[:invite_token]
+    @user = User.new
+  end
+
+  def create
+    @token = params[:user][:invite_token]
+
+    return if @token.nil?
+
+    @user = User.new(accept_params)
+    invite = Invite.find_by(token: @token)
+    @user.organization = invite.organization
+    @user.email = invite.email
+    @user.role = invite.role
+
+    if @user.save
+      invite.destroy
+      bypass_sign_in @user
+      flash[:success] = 'User successfully created.'
+      redirect_to account_path(@user)
+    else
+      flash[:danger] = 'Failed to create user.'
+      render action: :new, params: { invite_token: @token }
+    end
+  end
+
+  private
+
+  def accept_params
+    params.require(:user).permit(
+      :first_name,
+      :last_name,
+      :password,
+      :password_confirmation
+    )
+  end
+end
+
