+++ app/views/accept_invites/new.html.slim
@@ -1,0 +1,1 @@
+= render 'shared/sign_up_form', path: accept_invites_path, resource: @user, token: @token
