+++ app/views/devise/passwords/new.html.slim
@@ -11,7 +11,7 @@
                   h4.text-center.mb-2 Forgot Password?
               .form-group.d-flex.justify-content-center.align-items-center
                 = link_to 'Sign In', new_user_session_path, class: 'card-link btn btn-outline-primary text-nowrap'
-                = link_to 'Sign Up', registrations_path, class: 'card-link btn btn-outline-primary text-nowrap'
+                = link_to 'Sign Up', new_user_registration_path, class: 'card-link btn btn-outline-primary text-nowrap'
               .card-content.h-50.d-flex.align-items-center.mt-2
                 .card-body
                   .text-muted.text-center.mb-1
