+++ app/views/devise/registrations/new.html.slim
@@ -1,0 +1,1 @@
+= render 'shared/sign_up_form', path: user_registration_path, resource: @user
