+++ app/views/devise/shared/_sign_in.html.slim
@@ -31,6 +31,6 @@
   .text-center
     small.mr-25 Don't have an account?
     small
-      = link_to " Sign up", new_registration_path
+      = link_to " Sign up", new_user_registration_path(resource_name)
   br
