+++ app/views/shared/public/_header.html.slim
@@ -12,4 +12,4 @@
       = link_to "My Account", account_path, class: "btn btn-outline-primary"
     - else
       = link_to "Sign In", new_user_session_path, class: "btn btn-outline-primary"
-      = link_to "Sign Up", new_registration_path, class: "btn btn-outline-primary"
+      = link_to "Sign Up", new_user_registration_path, class: "btn btn-outline-primary"
